requirements.txt
Django~=5.0.6
django-crispy-forms~=2.1
crispy-bootstrap4~=2024.1
pillow~=10.3.0